refactor: change storageHlepers to helper
#110
